rename containers folder to pages
