+++ generators/app/index.js
@@ -114,15 +114,15 @@ class StackGenerator extends Generator {
       'client/src/components/Header/Header.container.js',
       'client/src/components/Header/logo.svg',
       'client/src/components/Header/style.css',
-      'client/src/containers/index.js',
-      'client/src/containers/Home/index.js',
-      'client/src/containers/Home/style.css',
-      'client/src/containers/Home/Home.component.js',
-      'client/src/containers/Home/Home.container.js',
-      'client/src/containers/Page/index.js',
-      'client/src/containers/Page/style.css',
-      'client/src/containers/Page/Page.component.js',
-      'client/src/containers/Page/Page.container.js',
+      'client/src/pages/index.js',
+      'client/src/pages/Home/index.js',
+      'client/src/pages/Home/style.css',
+      'client/src/pages/Home/Home.component.js',
+      'client/src/pages/Home/Home.container.js',
+      'client/src/pages/Page/index.js',
+      'client/src/pages/Page/style.css',
+      'client/src/pages/Page/Page.component.js',
+      'client/src/pages/Page/Page.container.js',
       'client/src/translations/fr.json',
       'client/src/utils/request.js',
       'client/src/utils/request.test.js',
