+++ generators/app/templates/client/src/pages/Home/Home.component.js
