+++ generators/app/templates/client/src/pages/Home/Home.container.js
