+++ generators/app/templates/client/src/pages/Home/index.js
