+++ generators/app/templates/client/src/pages/Home/style.css
