+++ generators/app/templates/client/src/pages/Page/Page.component.js
