+++ generators/app/templates/client/src/pages/Page/Page.container.js
