+++ generators/app/templates/client/src/pages/Page/index.js
