+++ generators/app/templates/client/src/pages/Page/style.css
