+++ generators/app/templates/client/src/pages/index.js
